Refactoring, new features: delimiter, parsing mode, multiprocessing
