+++ config.py
@@ -1,2 +1,3 @@
 DEFAULT_SCHEMA = 'dwh'
+DEFAULT_MODE = 'procedure'
 DELIMITER = ';;'
