+++ ezql.py
@@ -1,40 +1,41 @@
 import fire
 from utils.paths import *
-from antlrparser.tasks import FileProcessor
+from utils.logging import *
+from antlrparser.runner import Runner
 from output.json import to_json
 from output.cmd import beautify
+from config import *
 import os
-import logging
-
-logger = logging.getLogger(__name__)
-logger.setLevel(logging.DEBUG)
 
 
 class MyEzQl(object):
 
-    @staticmethod
-    def show(i: str, o: str=None) -> None:
-        """"""
+    @with_logging
+    def show(self, i: str, o: str='', ds='', dl=';;', p: bool=False) -> None:
 
         if o:
             if not (is_pathname_valid(o) and is_path_creatable(o) and o.endswith('.json')):
-                print(f'Error: {o} is whether not valid, or cannot be created')
+                logger.error(f'Error: {o} is whether not valid, or cannot be created')
                 return
 
-        processor = FileProcessor()
+        if not ds:
+            ds = DEFAULT_SCHEMA
+
+        p = 'procedure' if p else 'ddl'
+
+        processor = Runner(default_schema=ds, delimiter=dl, mode=p)
 
         if os.path.isdir(i):
             processor.parse_dir(i)
         elif os.path.isfile(i) and is_pathname_valid(i) and i.endswith('.sql'):
-            processor.parse_file(i)
+            processor.results = processor.parse_file(i)
         else:
-            print(f'Error: {i} is not a valid path')
+            logger.error(f'Error: {i} is not a valid path')
             return
-        r = processor.results
-        beautify(r)
+        beautify(processor.results)
         if o:
-            to_json(r, o)
-            print(f'-> {o} successfully saved')
+            to_json(processor.results, o)
+            logger.info(f'-> {o} successfully saved')
 
 
 if __name__ == '__main__':
