+++ output/cmd.py
@@ -15,20 +15,20 @@ def beautify(results: List[Procedure]) -> None:
 
     for p in results:
         msg += f'\n\n{Fore.GREEN}----------------------------\n'
-        msg += f'{Fore.GREEN}{p.name} ({p.path})\n'
+        msg += f'{Fore.GREEN}{p.schema}.{p.name} ({p.path})\n'
         msg += f'{Fore.GREEN}----------------------------\n{Style.RESET_ALL}'
         for q in p.queries:
             msg += f'|\n|--- {Fore.BLUE}{q.operation}{Style.RESET_ALL} ' \
-                   f'----> {Fore.CYAN}{q.target_table.name}\n{Style.RESET_ALL}'
+                   f'----> {Fore.CYAN}{q.target_table.schema}.{q.target_table.name}\n{Style.RESET_ALL}'
 
             if q.from_table:
                 for f in q.from_table:
                     msg += f'{Fore.CYAN}                  ' \
-                           f'. {Fore.BLUE}FROM{Fore.CYAN} {f.name}{Style.RESET_ALL}\n'
+                           f'. {Fore.BLUE}FROM{Fore.CYAN} {q.target_table.schema}.{f.name}{Style.RESET_ALL}\n'
             if q.join_table:
                 for f in q.join_table:
                     msg += f'{Fore.CYAN}                  ' \
-                           f'. {Fore.BLUE}JOIN{Fore.CYAN} {f.name}{Style.RESET_ALL}\n'
+                           f'. {Fore.BLUE}JOIN{Fore.CYAN} {q.target_table.schema}.{f.name}{Style.RESET_ALL}\n'
 
             if q.target_columns:
                 msg += f'. {Fore.BLUE}Columns --> {Fore.CYAN} {q.target_columns}{Style.RESET_ALL}\n\n'
