--- utils/extract.py
@@ -1,352 +1,0 @@
-from typing import Tuple, Optional, List
-from config.config import DEFAULT_SCHEMA
-from config.regex import *
-from antlr4 import ParserRuleContext, TerminalNode, ErrorNode, \
-    InputStream, CommonTokenStream
-from antlrparser.lexer import MySqlLexer
-from antlrparser.parser import MySqlParser
-import logging
-import re
-
-from sql.objects import Query, Table
-
-logger = logging.getLogger(__name__)
-logger.setLevel(logging.DEBUG)
-
-
-class Mapper:
-
-    """Helper class used to map DML extraction regexes with parsing
-    methods. Also provides cleanup regexes used to deal with edge
-    cases, e.g. table names containing SQL keywords"""
-
-    def __init__(self):
-
-        self.extract_regexes = \
-            {'INSERT': INSERT_REGEX,
-             'REPLACE': REPLACE_REGEX,
-             'UPDATE': UPDATE_REGEX,
-             'DELETE': DELETE_REGEX}
-
-        self.methods = None
-        self.parser = None
-        self.extractors = None
-        self.parsermethods = None
-        self.mapper = None
-
-    def map_methods(self):
-
-        """
-        Maps the required regexes and functions by DML type.
-
-        :param walker: Treewalker object
-        """
-
-        methods = \
-            {'INSERT': self.parser.insertStatement,
-             'REPLACE': self.parser.replaceStatement,
-             'UPDATE': self.parser.updateStatement,
-             'DELETE': self.parser.deleteStatement}
-
-        extractors = \
-            {'INSERT': parse_insert,
-             'REPLACE': parse_insert,
-             'UPDATE': parse_update,
-             'DELETE': parse_delete}
-
-        parsermethods = methods
-        self.mapper = {k: {'regex': v}
-                       for k, v in self.extract_regexes.items()}
-        for k in self.mapper.keys():
-            self.mapper[k]['parsermethod'] = parsermethods[k]
-            self.mapper[k]['extractor'] = extractors[k]
-
-
-def get_procedure_name(p: str) -> Tuple[str, str]:
-
-    """
-    Gets the procedure name from the procedure body.
-
-    :param p: procedure string
-    :return: tuple (schema_string, name_string)
-    """
-
-    name = re.search(NAME_REGEX, p).group().lower()
-    schema, name = parse_object_name(name)
-    return schema, name
-
-
-def parse_object_name(name: str, default_schema: str=DEFAULT_SCHEMA) -> Tuple[str, str]:
-
-    """
-    Separates name and schema from a raw database object name.
-    If no schema is found, uses the default schema defined
-    in the configuration.
-
-    :param name: raw table name
-    :param default_schema: default schema, defaults to config file definition
-    :return: tuple (schema_string, name_string)
-    """
-
-    if '.' in name:
-        schema = name.split('.')[0]
-        name = name.split('.')[1]
-    else:
-        schema = default_schema
-
-    return schema, name
-
-
-def parse_statement(dmltype: str, s: str, mapper: Mapper) -> Query:
-
-    """
-    Cleans DML statement, creates parsing objects, and returns a
-    Query object.
-
-    :param dmltype: type of DML statement being parsed
-    :param s: DML statement string
-    :param mapper: Mapper object
-    :return: Query object containing the statement information
-    """
-
-    input = InputStream(s)
-    lexer = MySqlLexer(input)
-    stream = CommonTokenStream(lexer)
-    parser = MySqlParser(stream)
-    mapper.parser = parser
-    mapper.map_methods()
-    tree = mapper.mapper[dmltype]['parsermethod']()
-    r = mapper.mapper[dmltype]['extractor'](tree, dmltype)
-    return r
-
-
-def get_updated_columns(tree: ParserRuleContext) -> List[str]:
-
-    """
-    Loops recursively over AST children and gathers all the updated
-    columns.
-
-    :param tree: AST object parsed
-    :return: list containing all update details
-    """
-
-    cols = []
-    for c in tree.getChildren():
-
-        if isinstance(c, MySqlParser.UpdatedElementContext):
-            cols.append(c.getText().lower())
-        elif not isinstance(c, TerminalNode):
-            cols.extend(get_updated_columns(c))
-
-    return cols
-
-
-def get_inserted_columns(tree: ParserRuleContext) -> List[str]:
-
-    """
-    Loops recursively over AST children and gathers all the inserted
-    columns.
-
-    :param tree: AST object parsed
-    :return: list of columns being inserted into
-    """
-
-    for child in tree.getChildren():
-        if isinstance(child, MySqlParser.UidListContext):
-            target_columns = child.getText().lower().split(',')
-            return target_columns
-
-    target_columns = []
-    return target_columns
-
-
-def parse_update(tree: ParserRuleContext, dmltype: str) \
-        -> Optional[Query]:
-
-    """
-    Parses target table, source table(s) and target columns from
-    an UPDATE statement.
-
-    :param tree: AST object parsed
-    :param dmltype: DML statement type is passed for Query object
-    instantiation
-    :return: Query object containing the statement information
-    """
-
-    q = Query(operation=dmltype)
-
-    q.target_table = get_target_table(tree)
-    q.join_table = get_source_tables_update(tree)
-    q.target_columns = get_updated_columns(tree)
-
-    if q.target_table and q.join_table:
-        msg = {'join': [x.name for x in q.join_table if x]}
-        logger.info(f'UPDATE: {msg}')
-        return q
-
-
-def get_inserted_tables(tree: ParserRuleContext) \
-        -> Tuple[List[Table], List[Table]]:
-
-    """
-    Gets all source tables from an INSERT statement.
-
-    :param tree: AST object parsed
-    :return: list of source tables names string
-    """
-
-    for child in tree.getChildren():
-        if isinstance(child, MySqlParser.InsertStatementValueContext):
-            from_table = get_source_tables_insert(child, 'from')
-            join_table = get_source_tables_insert(child, 'join')
-            msg = {'from': [x.name for x in from_table if x],
-                   'join': [x.name for x in join_table if x]}
-            logger.info(f'INSERT/REPLACE: {msg}')
-            return from_table, join_table
-
-def parse_insert(tree: ParserRuleContext, dmltype: str) -> Query:
-
-    """
-    Parses target table, source table(s) and target columns from
-    an INSERT statement.
-
-    :param tree: AST object parsed
-    :param dmltype: DML statement type for Query object instantiation
-    :return: Query object containing the statement information
-    """
-
-    q = Query(operation=dmltype)
-    q.target_table = get_target_table(tree)
-    q.from_table, q.join_table = get_inserted_tables(tree)
-    q.target_columns = get_inserted_columns(tree)
-
-    return q
-
-
-def get_delete_table(tree: ParserRuleContext) \
-        -> Optional[List[Table]]:
-
-    """
-    Gets the table name from a DELETE statement.
-
-    :param tree: AST object parsed
-    :return: the target table name
-    """
-
-    for child in tree.getChildren():
-        if isinstance(child, MySqlParser.DeleteStatementValueContext):
-            return get_source_tables_insert(child, 'from')
-
-
-def parse_delete(tree: ParserRuleContext, dmltype: str) -> Query:
-
-    """
-    Parses the target table from a DELETE statement
-    :param tree: AST object parsed
-    :param dmltype: DML statement type for Query object instantiation
-    :return: Query object containing the statement information
-    """
-
-    q = Query(operation=dmltype)
-    q.target_table = get_target_table(tree)
-
-    return q
-
-
-def get_source_tables_update(tree: ParserRuleContext) \
-        -> List[Table]:
-
-    """
-    Collects recursively table names in JOIN clauses of an
-    update statement.
-
-    :param tree: AST object parsed
-    :return: list of source tables
-    """
-
-    tables = []
-
-    for c in tree.getChildren():
-        if isinstance(c, MySqlParser.JoinPartContext):
-            tables.extend(get_all_tables(c))
-        elif not (isinstance(c, TerminalNode) or isinstance(c, ErrorNode)):
-            tables.extend(get_source_tables_update(c))
-
-    return tables
-
-
-def get_source_tables_insert(tree: ParserRuleContext, clause: str) \
-        -> List[Table]:
-
-    """
-    Collects recursively table names in FROM and JOIN clauses in
-    insert/replace statement.
-
-    :param tree: AST object parsed
-    :param clause: clause type, must be one of ('from', 'join')
-    :return: list of source tables
-    """
-
-    assert clause in ('from', 'join'), f'variable clause must be one of ' \
-                                       f'(from, join)'
-
-    tables = []
-
-    for c in tree.getChildren():
-        if isinstance(c, MySqlParser.FromClauseContext):
-            froms = c.tableSources()
-            for i in froms.tableSource():
-                if clause == 'from':
-                    f = i.tableSourceItem()
-                    tables.extend(get_all_tables(f))
-                else:
-                    f = i.joinPart()
-                    for x in f:
-                        tables.extend(get_all_tables(x))
-        elif not (isinstance(c, TerminalNode) or isinstance(c, ErrorNode)):
-            tables.extend(get_source_tables_insert(c, clause))
-    return tables
-
-
-def get_target_table(tree: ParserRuleContext) -> Table:
-
-    """
-    Walks recursively to the first table name found in a statement AST
-    and returns it.
-
-    :param tree: AST object parsed
-    :return: Table object
-    """
-
-    for c in tree.getChildren():
-
-        if isinstance(c, MySqlParser.TableNameContext):
-            schema, name = parse_object_name(c.getText().lower())
-            t = Table(name, schema)
-            return t
-        elif not (isinstance(c, TerminalNode) or isinstance(c, ErrorNode)):
-            return get_target_table(c)
-
-
-def get_all_tables(tree: ParserRuleContext) -> List[Table]:
-
-    """
-    Gets all table names inside a tree, appends them to a list of
-    table names and finally returns the list.
-
-    :param tree: AST object parsed
-    :return: list of table names
-    """
-
-    tables = []
-    for c in tree.getChildren():
-        if isinstance(c, MySqlParser.TableNameContext):
-            schema, name = parse_object_name(c.getText().lower())
-            t = Table(name, schema)
-            tables.append(t)
-        elif isinstance(c, MySqlParser.QueryExpressionContext):
-            c = c.querySpecification()
-            tables.extend(get_all_tables(c))
-        elif not (isinstance(c, TerminalNode) or isinstance(c, ErrorNode)):
-            tables.extend(get_all_tables(c))
-    return tables
